requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
supabase==2.1.0
stripe==6.5.0
Flask-Login==0.6.2
pdfplumber==0.10.2
python-docx==0.8.11
Werkzeug==2.3.7
gunicorn==21.2.0
python-magic==0.4.27
PyPDF2==3.0.1
pdf2image==1.16.3
Pillow==10.0.0
beautifulsoup4==4.12.2
pytesseract==0.3.10
docx2pdf==0.1.8
pymupdf==1.23.7 
reportlab==4.4.1